+++ src/Sweeper.sol
@@ -23,13 +23,13 @@ contract Sweeper {
         _token.transfer(_to, _value);
     }
 
-    function sweepAll(
-        address[] calldata _tokens,
-        address _to
-    ) public onlyOwner {
-        for (uint i = 0; i < _tokens.length; i++) {
+    function sweepAll(address[] calldata _tokens, address _to)
+        public
+        onlyOwner
+    {
+        for (uint256 i = 0; i < _tokens.length; i++) {
             IERC20 token = IERC20(_tokens[i]);
-            uint balance = token.balanceOf(address(this));
+            uint256 balance = token.balanceOf(address(this));
             token.transfer(_to, balance);
         }
     }
